Combined the scripts and move preproc_kc.py
